refactor: move codes to its own file
