+++ lib/candevice.js
@@ -18,7 +18,8 @@ const debug = require('debug')('canboatjs:candevice')
 const EventEmitter = require('events')
 const _ = require('lodash')
 const Uint64LE = require('int64-buffer').Uint64LE
-const { getManufacturerCode, getIndustryCode, getDeviceClassCode, getCanIdFromPGN, defaultTransmitPGNs } = require('./utilities')
+const { getCanIdFromPGN } = require('./utilities')
+const { defaultTransmitPGNs, getIndustryCode, getManufacturerCode, getDeviceClassCode } = require('./codes')
 const { toPgn } = require('./toPgn')
 
 const addressClaim = {
@@ -35,7 +36,6 @@ const addressClaim = {
   "Reserved1": 1,
   "Reserved2": 2
 }
-
 
 class CanDevice extends EventEmitter {
   constructor (canbus, options) {
