+++ lib/fromPgn.js
@@ -26,6 +26,7 @@ const Int64LE = require('int64-buffer').Int64LE
 const Uint64LE = require('int64-buffer').Uint64LE
 const { parse: parseDate } = require('date-fns')
 const { getPGNFromCanId, getManufacturerName, getIndustryName } = require('./utilities')
+const { getManufacturerName } = require('./codes')
 
 var fieldTypeReaders = {}
 var fieldTypePostProcessors = {}
@@ -246,7 +247,7 @@ class Parser extends EventEmitter {
     } catch ( error ) {
       this.emit('error', pgn, error)
       cb && cb(error)
-      return 
+      return
     }
   }
 
@@ -282,7 +283,7 @@ class Parser extends EventEmitter {
   isN2KOver0183(sentence) {
     return sentence.startsWith('$PCDIN,')
   }
-  
+
   parseN2KOver0183(sentence, cb) {
     if ( sentence.startsWith('$PCDIN,') ) {
       return this.parsePCDIN(sentence, cb)
@@ -334,14 +335,14 @@ class Parser extends EventEmitter {
     pgn.src = parseInt(split[3], 16)
     pgn.dst = 255
     pgn.prio = 0
-    
+
     var data = split[4]
     if ( data.indexOf('*') != -1 ) {
       data = data.split('*')[0]
     }
-    
+
     const bs = new BitStream(Buffer.from(data, 'hex'))
-    
+
     const res = this._parse(pgn, bs, data.length/2, cb)
     if ( res ) {
       debug('parsed pgn %j', pgn)
@@ -355,7 +356,7 @@ class Parser extends EventEmitter {
       if ( this.isN2KOver0183(pgn_data) ) {
         return this.parseN2KOver0183(pgn_data, cb)
       }
-      
+
       var split = pgn_data.split(',')
 
       var pgn = {}
