+++ lib/ikonvert.js
@@ -24,7 +24,8 @@ const Parser = require('./fromPgn').Parser
 const _ = require('lodash')
 const CanDevice = require('./candevice')
 const spawn = require('child_process').spawn
-const { getPGNFromCanId, getCanIdFromPGN, actisenseSerialToBuffer, defaultTransmitPGNs } = require('./utilities')
+const { getPGNFromCanId, getCanIdFromPGN, actisenseSerialToBuffer } = require('./utilities')
+const { defaultTransmitPGNs } = require('./codes')
 
 const pgnsSent = {}
 const rateLimit = 200
