+++ lib/toPgn.js
@@ -19,7 +19,8 @@ const _ = require('lodash')
 const BitStream = require('bit-buffer').BitStream
 const Int64LE = require('int64-buffer').Int64LE
 const Uint64LE = require('int64-buffer').Uint64LE
-const { getPlainPGNs, getManufacturerCode, getIndustryCode, getCanIdFromYdgwPGN , actisenseSerialToBuffer} = require('./utilities')
+const { getPlainPGNs, getCanIdFromYdgwPGN , actisenseSerialToBuffer} = require('./utilities')
+const { getIndustryCode, getManufacturerCode } = require('./codes')
 
 const RES_STRINGLAU = 'ASCII or UNICODE string starting with length and control byte'
 
@@ -340,7 +341,7 @@ fieldTypeWriters['ASCII text'] = (pgn, field, value, bs) => {
       }
     }
   }
-  
+
   if ( _.isUndefined(value) ) {
     value = ""
   }
