+++ lib/ydgw02.js
@@ -19,7 +19,8 @@ const Transform = require('stream').Transform
 const FromPgn = require('./fromPgn').Parser
 const Parser = require('./fromPgn').Parser
 const _ = require('lodash')
-const { getCanIdFromYdgwPGN, getPGNFromCanId, actisenseSerialToBuffer, defaultTransmitPGNs } = require('./utilities')
+const { getCanIdFromYdgwPGN, getPGNFromCanId, actisenseSerialToBuffer } = require('./utilities')
+const { defaultTransmitPGNs } = require('./codes')
 const { pgnToYdgwRawFormat, actisenseToYdgwRawFormat } = require('./toPgn')
 
 const pgnsSent = {}
@@ -86,7 +87,7 @@ Ydgw02Stream.prototype.sendYdgwPGN = function (msg) {
   actisenseToYdgwRawFormat(msg).forEach(raw => {
     this.sendString(raw)
   })
-  
+
   /*
   if ( !this.parser ) {
     this.parser = new Parser()
@@ -97,7 +98,7 @@ Ydgw02Stream.prototype.sendYdgwPGN = function (msg) {
       console.error(error.stack)
     })
 
-    
+
     this.parser.on('pgn', (pgn) => {
       let now = Date.now()
       let lastSent = pgnsSent[pgn.pgn]
